Fixed serialization of PlanarSurfaces
