+++ openquake/calculators/ucerf_risk.py
@@ -68,9 +68,9 @@ def compute_ruptures(sources, sitecol, gsims, monitor):
                 eid += 1
             if events:
                 ebruptures.append(
-                    event_based.EBRupture(
+                    calc.EBRupture(
                         rup, indices,
-                        numpy.array(events, event_based.event_dt),
+                        numpy.array(events, calc.event_dt),
                         src.source_id, src.src_group_id, serial))
                 serial += 1
                 res.num_events += len(events)
